fourth try

+++ Filling Jars/main.cpp
@@ -7,11 +7,11 @@ using namespace std;
 int main() {
     int n, m, a, b, k;
     cin >> n >> m;
-	long long av = 0;
+	unsigned __int128 av = 0;
     for ( int i=0; i<m; ++i ) {
         cin >> a >> b >> k;
         av += (k*(b-a+1));
     }
-    cout << av/n;
+    cout << (long long)av/n;
     return 0;
 }
